Finished converting tools directory, cleaned up more unused code
