+++ gunfolds/old/unused.py
@@ -1,6 +1,6 @@
 """ This is where functions from old experiments and previous data formats have been moved
     to in order to clarify the current code.  These remain for reference and in case they
-    become useful again.  No garuntee all imports are here to use these functions."""
+    become useful again.  No guarantee all imports are here to use these functions."""
 import numpy as np
 import scipy
 
@@ -1225,3 +1225,758 @@ def unroll_undersample(G, steps):
         N.update(
             {v: set([nodes[k] for k in scipy.where([ecj.reachable(v + '0', U, u + str(steps - 1)) for u in G])[0]])})
     return N
+
+
+
+
+
+
+
+
+
+
+""" from unknownrate.py """
+
+
+def eqclass(H):
+    '''
+    Find all graphs in the same equivalence class with respect to
+    graph H and any undesampling rate.
+    '''
+    g = {n: {} for n in H}
+    s = set()
+
+    @memo
+    def addedges(g, H, edges):
+        if edges:
+            nedges = prune_conflicts(H, g, edges)
+            n = len(nedges)
+
+            if n == 0:
+                return None
+
+            for i in range(n):
+                gk.addanedge(g, nedges[i])
+                if bfu.call_u_equals(g, H):
+                    s.add(g2num(g))
+                addedges(g, H, nedges[:i] + nedges[i+1:])
+                gk.delanedge(g, nedges[i])
+    edges = gk.edgelist(gk.complement(g))
+    addedges(g, H, edges)
+
+    return s - set([None])
+
+# these two functions come from this answer:
+# http://stackoverflow.com/a/12174125
+
+def set_bit(value, bit):
+    return value | (1 << bit)
+
+
+def clear_bit(value, bit):
+    return value & ~(1 << bit)
+
+
+def hashloop(l):
+    t = [int(x) for x in l]
+    idx = np.argmin(t)
+    return tuple(l[idx:] + l[:idx])
+
+
+def perm_circular_slow2(l):
+    s = [tuple(l)]
+    c = {}
+    c[hashloop(l)] = True
+    for e in permutations(l):
+        if not hashloop(e) in c:
+            s.append(e)
+            c[hashloop(e)] = True
+    return s
+
+
+def perm_cyclic(l):
+    return [tuple(l[i:] + l[:i]) for i in range(len(l))]
+
+
+def perm_circular_slow(l):
+    s = [tuple(l)]
+    c = set(perm_cyclic(l))
+    for e in permutations(l):
+        if not e in c:
+            s.append(e)
+            c = c | set(perm_cyclic(e))
+    return s
+
+
+def may_be_true_selfloop(n, H):
+    for v in H[n]:
+        if v == n:
+            continue
+        if H[n][v] == 1:
+            return False
+    return True
+
+
+def issingleloop(num):
+    bl = gmp.bit_length(num)
+    idx = [1 for i in xrange(bl) if num & (1 << i)]
+    return len(idx) == 1
+
+
+def nonbarren(H):
+    for v in H:
+        if H[v]:
+            return v
+    return False
+
+
+def count_loops(n):
+    s = 0
+    for i in range(1, n + 1):
+        s += comb(n, i) * math.factorial(i - 1)
+    return s
+
+
+def noverlap_loops(loops):
+    d = {}
+    for l in loops:
+        el = []
+        for k in loops:
+            if not set(l) & set(k):
+                el.append(tuple(k))
+                # d.setdefault(tuple(l),set()).add(tuple(k))
+        d[tuple(l)] = noverlap_loops(el)
+    return d
+
+
+def loop_combinations(loops):
+    s = set()
+    d = noverlap_loops(loops)
+
+    def dfs_traverse(d, gs):
+        if d:
+            for e in d:
+                dfs_traverse(d[e], gs | set([e]))
+        else:
+            s.add(frozenset(gs))
+    for e in d:
+        dfs_traverse(d[e], set([e]))
+    return list(s)
+
+
+def sorted_loops(g):
+    l = [x for x in sl.simple_loops(g, 0)]
+    s = {}
+    for e in l:
+        s.setdefault(len(e), []).append(e)
+    return s
+
+
+def loopgroups(g, n=None):
+    d = sorted_loops(g)
+    if n:
+        return loop_combinations(d[n])
+    else:
+        l = []
+        for key in d:
+            l.append(loop_combinations(d[key]))
+        return l
+
+
+def rotate(l, n):
+    return l[n:] + l[:n]
+
+
+def get_perm(loop1, loop2, n=None):
+    if not n:
+        n = len(loop1)
+    basel = [i for i in xrange(1, n + 1)]
+    diff1 = set(basel) - set(loop1)
+    diff2 = set(basel) - set(loop2)
+    if loop1[0] in loop2:
+        l2 = rotate(loop2, loop2.index(loop1[0]))
+    else:
+        l2 = loop2
+    mp = {}
+    for x, y in zip(loop1 + list(diff1), l2+list(diff2)):
+        mp[x] = y
+    return mp
+
+
+def permute(g, perm):
+    gn = {x: {} for x in g}
+    for e in g:
+        gn[perm[e]] = {perm[x]: g[e][x] for x in g[e]}
+    return gn
+
+
+def permuteAset(s, perm):
+    n = len(perm)
+    ns = set()
+    for e in s:
+        ns.add(g2num(permute(num2CG(e, n), perm)))
+    return ns
+
+
+def set_loop(loop, graph):
+    for i in range(0, len(loop) - 1):
+        graph[loop[i]][loop[i + 1]] = 1
+    graph[loop[-1]][loop[0]] = 1
+
+
+def add2set(gset, elist, H):
+    n = len(H)
+
+    s = set()
+    ss = set()
+
+    eremove = {e: True for e in elist}
+
+    for gnum in gset:
+        g = num2CG(gnum, n)
+        for e in elist:
+            if not e[1] in g[e[0]]:
+                gk.addanedge(g, e)
+                num = g2num(g)
+                if not num in s:
+                    au = bfu.call_undersamples(g)
+                    if not bfu.check_conflict(H, g, au=au):
+                        eremove[e] = False
+                        s.add(num)
+                        if bfu.check_equality(H, g, au=au):
+                            ss.add(num)
+                gk.delanedge(g, e)
+
+    for e in eremove:
+        if eremove[e]:
+            elist.remove(e)
+
+    return s, ss, elist
+
+
+def eqclass_list(H):
+    '''
+    Find all graphs in the same equivalence class with respect to
+    graph H and any undesampling rate.
+    '''
+    g = {n: {} for n in H}
+    s = set()
+
+    edges = gk.edgelist(gk.complement(g))
+    # edges = prune_conflicts(H, g, edges)
+
+    gset = set([g2num(g)])
+    for i in range(len(H) ** 2):
+        print i
+        gset, ss, edges = add2set(gset, edges, H)
+        s = s | ss
+        if not edges:
+            break
+
+    return s
+
+
+def quadmerge_(glist, H, ds):
+    n = len(H)
+    gl = set()
+    ss = set()
+    conflicts = set()
+    for gi in combinations(glist, 2):
+        if gi[0] & gi[1]:
+            continue
+        # if skip_conflict(gi[0], gi[1], ds): continue
+        gnum = gi[0] | gi[1]
+        if gnum in conflicts:
+            continue
+        if skip_conflictors(gnum, ds):
+            conflicts.add(gnum)
+            continue
+        if gnum in gl:
+            continue
+        g = num2CG(gnum, n)
+        if not bfu.call_u_conflicts(g, H):
+            gl.add(gnum)
+            if bfu.call_u_equals(g, H):
+                ss.add(gnum)
+        else:
+            conflicts.add(gnum)
+    return gl, ss
+
+
+def edgemask(gl, H, cds):
+    """given a list of encoded graphs and observed undersampled graph
+    H returns a matrix with -1 on diagonal, 0 at the conflicting graph
+    combination and encoded graph at non-conflicted
+    positions. Furthermore, returns a set of graphs that are in the
+    equivalence class of H
+
+    Arguments:
+    - `gl`: list of integer encoded graphs
+    - `H`: the observed undersampled graph
+    """
+    n = len(H)
+    nl = len(gl)
+    s = set()
+    mask = np.zeros((nl, nl), 'int')
+    np.fill_diagonal(mask, -1)
+
+    for i in xrange(nl):
+        for j in xrange(i + 1, nl):
+
+            if gl[i] & gl[j]:
+                continue
+            if skip_conflict(gl[i], gl[j], cds):
+                continue
+
+            gnum = gl[i] | gl[j]
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+                mask[i, j] = gnum
+                mask[j, i] = gnum
+    return mask, s
+
+
+def ledgemask(gl, H, cds):
+    """given a list of encoded graphs and observed undersampled graph
+    H returns a matrix with -1 on diagonal, 0 at the conflicting graph
+    combination and encoded graph at non-conflicted
+    positions. Furthermore, returns a set of graphs that are in the
+    equivalence class of H
+
+    Arguments:
+    - `gl`: list of integer encoded graphs
+    - `H`: the observed undersampled graph
+    """
+    n = len(H)
+    nl = len(gl)
+    s = set()
+    mask = np.zeros((nl, nl), 'int')
+    np.fill_diagonal(mask, -1)
+
+    for i in xrange(nl):
+        for j in xrange(i + 1, nl):
+
+            if gl[i] & gl[j]:
+                continue
+            gnum = gl[i] | gl[j]
+            if skip_conflictors(gnum, cds):
+                continue
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+                mask[i, j] = gnum
+                mask[j, i] = gnum
+    return mask, s
+
+
+def edgeds(mask):
+    """construct an edge dictionary from the mask matrix
+
+    Arguments:
+    - `mask`:
+    """
+    ds = {}
+    nl = mask.shape[0]
+    idx = np.triu_indices(nl, 1)
+    for i, j in zip(idx[0], idx[1]):
+        if mask[i, j]:
+            ds[(i, j)] = set()
+            conf = set([i, j])
+            conf = conf.union(np.where(mask[i,:] == 0)[0])
+            conf = conf.union(np.where(mask[j,:] == 0)[0])
+            for k, m in zip(idx[0], idx[1]):
+                if not mask[k, m]:
+                    continue
+                if k in conf:
+                    continue
+                if m in conf:
+                    continue
+                if not (k, m) in ds:
+                    ds[(i, j)].add(mask[k, m])
+            if not ds[(i, j)]:
+                ds.pop((i, j))
+    return ds
+
+
+def edgedsg(mask):
+    """construct an edge dictionary from the mask matrix
+
+    Arguments:
+    - `mask`:
+    """
+    ds = {}
+    nl = mask.shape[0]
+    idx = np.triu_indices(nl, 1)
+    for i, j in zip(idx[0], idx[1]):
+        if mask[i, j]:
+            ds[mask[i, j]] = set()
+            conf = set([i, j])
+            conf = conf.union(np.where(mask[i,:] == 0)[0])
+            conf = conf.union(np.where(mask[j,:] == 0)[0])
+            for k, m in zip(idx[0], idx[1]):
+                if not mask[k, m]:
+                    continue
+                if k in conf:
+                    continue
+                if m in conf:
+                    continue
+                if not (k, m) in ds:
+                    ds[mask[i, j]].add(mask[k, m])
+            if not ds[mask[i, j]]:
+                ds.pop(mask[i, j])
+    return ds
+
+
+def quadlister(glist, H, cds):
+    n = len(H)
+    s = set()
+    cache = {}
+
+    def edgemask(gl, H, cds):
+        nl = len(gl)
+        ss = set()
+        mask = np.zeros((nl, nl), 'int')
+        np.fill_diagonal(mask, -1)
+        idx = np.triu_indices(nl, 1)
+        for i, j in zip(idx[0], idx[1]):
+            if gl[i] & gl[j]:
+                mask[i, j] = -1
+                mask[j, i] = -1
+                continue
+            if skip_conflict(gl[i], gl[j], cds):
+                gnum = gl[i] | gl[j]
+                cache[gnum] = False
+                continue
+
+            gnum = gl[i] | gl[j]
+            if gnum in cache:
+                if cache[gnum]:
+                    mask[i, j] = gnum
+                    mask[j, i] = gnum
+            else:
+                cache[gnum] = False
+                g = num2CG(gnum, n)
+                if not bfu.call_u_conflicts(g, H):
+                    if bfu.call_u_equals(g, H):
+                        ss.add(gnum)
+                    mask[i, j] = gnum
+                    mask[j, i] = gnum
+                    cache[gnum] = True
+        return mask, ss
+
+    def quadmerger(gl, H, cds):
+        mask, ss = edgemask(gl, H, cds)
+        ds = edgeds(mask)
+        # ipdb.set_trace()
+        return [[mask[x]] + list(ds[x]) for x in ds], ss
+
+    l = []
+    for gl in glist:
+        ll, ss = quadmerger(gl, H, cds)
+        l.extend(ll)
+        s = s | ss
+
+    return l, s
+
+
+def dceqc(H):
+    """Find all graphs in the same equivalence class with respect to H
+
+    Arguments:
+    - `H`: an undersampled graph
+    """
+    if bfu.is_sclique(H):
+        print 'not running on superclique'
+        return set()
+    n = len(H)
+    s = set()
+    cds = confpairs(H)
+
+    glist =  [2 ** np.arange(n**2)]
+    i = 1
+    # for i in range(int(np.log2(n**2))):
+    while glist != []:
+        print i, np.max(map(len, glist)), len(glist)
+        glist, ss = quadlister(glist, H, cds)
+        s = s | ss
+        i += 1
+    return s
+
+
+def quadmerge(gl, H, cds):
+    n = len(H)
+    l = set()
+    s = set()
+    mask, ss = edgemask(gl, H, cds)
+    s = s | ss
+    ds = edgeds(mask)
+
+    # pp = pprint.PrettyPrinter(indent=1)
+    # pp.pprint(ds)
+
+    for idx in ds:
+        for gn in ds[idx]:
+            if mask[idx] & gn:
+                continue
+            if skip_conflict(mask[idx], gn, cds):
+                continue
+            gnum = mask[idx] | gn
+            if gnum in l or gnum in ss:
+                continue
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                l.add(gnum)
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+
+    return list(l), s
+
+
+def edgemask2(gl, H, cds):
+    n = len(H)
+    nl = len(gl)
+    s = set()
+    o = set()
+    mask = np.zeros((nl, nl), 'int')
+    np.fill_diagonal(mask, -1)
+    for i in xrange(nl):
+        for j in xrange(i + 1, nl):
+            if gl[i] & gl[j]:
+                continue
+            gnum = gl[i] | gl[j]
+            if skip_conflictors(gnum, cds):
+                continue
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+                mask[i, j] = gnum
+                mask[j, i] = gnum
+            elif bfu.overshoot(g, H):
+                o.add(gnum)
+    return mask, s, o  # mask, found eqc members, overshoots
+
+
+def ecmerge(H):
+    """Find all graphs in the same equivalence class with respect to H
+
+    Arguments:
+    - `H`: an undersampled graph
+    """
+    if bfu.is_sclique(H):
+        print 'not running on superclique'
+        return None
+    n = len(H)
+    s = set()
+    ds = confpairs(H)
+    ccf = conflictors(H)
+    cset = set()
+    for e in ccf:
+        cset = cset.union(ccf[e])
+
+    glist =  np.r_[[0], 2 ** np.arange(n**2)]
+    # glist =  2**np.arange(n**2)
+
+    # glist, ss = quadmerge(glist,H)
+
+    for i in range(int(2 * np.log2(n))):
+        print i, len(glist)
+        glist, ss = quadmerge_(glist, H, cset)
+        s = s | ss
+    return s
+
+
+def getrates(g, H):
+    n = len(H)
+    au = bfu.call_undersamples(g)
+    return list(np.where(map(lambda x: x == H, au))[0])
+
+
+def withrates(s, H):
+    n = len(H)
+    d = {g: set() for g in s}
+    for g in s:
+        d[g] = getrates(num2CG(g, n), H)
+    return d
+
+
+def patchmerge(ds, H, cds):
+    n = len(H)
+    l = set()
+    s = set()
+    o = set()
+    for gkey in ds:
+        for num in ds[gkey]:
+            if gkey & num:
+                continue
+            gnum = gkey | num
+            if gnum is s:
+                continue
+            if skip_conflictors(gnum, cds):
+                continue
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                l.add(gnum)
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+            elif not gnum in o and bfu.overshoot(g, H):
+                o.add(gnum)
+    return l, s, o
+
+
+def quadmerge2(gl, H, cds):
+    n = len(H)
+
+    mask, s, o = edgemask2(gl, H, cds)
+    # ipdb.set_trace()
+    ds = edgedsg(mask)
+    l, ss, oo = patchmerge(ds, H, cds)
+
+    o = o | oo
+    s = s | ss
+
+    print 'overshoots: ', len(o)
+
+    return list(l), s
+
+
+def quadmerge21(gl, H, cds):
+    n = len(H)
+    l = set()
+
+    mask, ss, o = edgemask2(gl, H, cds)
+    idx = np.triu_indices(mask.shape[0], 1)
+    print len(o)
+    for i in range(len(idx[0])):
+        if mask[idx[0][i], idx[1][i]]:
+            l.add(mask[idx[0][i], idx[1][i]])
+
+    return list(l), ss
+
+
+def dceqclass2(H):
+    """Find all graphs in the same equivalence class with respect to H
+
+    Arguments:
+    - `H`: an undersampled graph
+    """
+    if bfu.is_sclique(H):
+        print 'not running on superclique'
+        return set()
+    n = len(H)
+    s = set()
+    cp = confpairs(H)
+    confs = conflictor_set(H)
+    ccf = conflictors(H)
+
+    def prune_loops(gl, H):
+        l = []
+        for e in gl:
+            if e[0] == e[1] and not (e[1] in H[e[0]] and H[e[0]][e[1]] in (1,3)):
+                continue
+            l.append(e)
+        return l
+    edges = gk.edgelist(gk.complement(num2CG(0, n)))
+    edges = prune_loops(edges, H)
+    glist = map(lambda x: e2num(x, n), edges)
+
+    # glist =  list(2**np.arange(n**2))
+    i = 0
+    while glist != []:
+        print 2 ** i, len(glist)
+        glist_prev = glist
+        glist, ss = quadmerge21(glist, H, confs)
+        s = s | ss
+        i += 1
+
+    ds = {x: edges for x in glist_prev}
+
+    for j in range(i, len(H) ** 2):
+        ds, ss = add2set_(ds, H, cp, ccf, iter=j, verbose=True)
+        s = s | ss
+        if not ds:
+            break
+
+    return s
+
+
+def dceqclass(H):
+    """Find all graphs in the same equivalence class with respect to H
+
+    Arguments:
+    - `H`: an undersampled graph
+    """
+    if bfu.is_sclique(H):
+        print 'not running on superclique'
+        return set()
+    n = len(H)
+    s = set()
+    cds = confpairs(H)
+
+    glist =  [0] + list(2**np.arange(n**2))
+    i = 1
+    while glist != []:
+        print i, len(glist)
+        glist, ss = quadmerge(glist, H, cds)
+        s = s | ss
+        i += 1
+    return s
+
+
+def ldceqclass(H, asl=None):
+    """Find all graphs in the same equivalence class with respect to H
+
+    Arguments:
+    - `H`: an undersampled graph
+    """
+    if bfu.is_sclique(H):
+        print 'not running on superclique'
+        return set()
+    n = len(H)
+    s = set()
+    cds = lconfpairs(H)
+    if asl:
+        sloops = asl
+    else:
+        sloops = prune_loops(allsloops(len(H)), H)
+
+    glist = sloops
+    i = 1
+    while glist != []:
+        print i, len(glist)
+        glist, ss = lquadmerge(glist, H, cds)
+        s = s | ss
+        i += 1
+    return s
+
+
+def lquadmerge(gl, H, cds):
+    n = len(H)
+    l = set()
+    s = set()
+    mask, ss = ledgemask(gl, H, cds)
+    s = s | ss
+    ds = edgeds(mask)
+
+    # pp = pprint.PrettyPrinter(indent=1)
+    # pp.pprint(ds)
+
+    for idx in ds:
+        for gn in ds[idx]:
+            if mask[idx] & gn:
+                continue
+            if skip_conflictors(mask[idx], gn, cds):
+                continue
+            gnum = mask[idx] | gn
+            if gnum in l or gnum in ss:
+                continue
+            g = num2CG(gnum, n)
+            if not bfu.call_u_conflicts(g, H):
+                l.add(gnum)
+                if bfu.call_u_equals(g, H):
+                    s.add(gnum)
+
+    return list(l), s
