+++ gunfolds/tools/simpleloops.py
@@ -5,12 +5,6 @@ from numpy import argsort
 
 
 
-
-##++++ CONVERTED ++++##
-
-
-
-
 def simple_loops(g, u):
     """
     iterator over the list of simple loops of graph g at the undersample rate u
