+++ tests/tests.py
@@ -199,6 +199,55 @@ class TestUnknownRateFunctions(unittest.TestCase):
         s = ur.liteqclass(g2, verbose=False, capsize=1000)
         self.assertEqual(s, set_of_u_2)
 
+    def test__gen_loops(self):
+        n = 3
+        expected = [[1], [2], [3], [1, 2], [1, 3],
+                    [2, 3], [1, 2, 3], [1, 3, 2]]
+        self.assertEqual(expected, ur.gen_loops(n))
+
+        n = 5
+        expected = [[1], [2], [3], [4], [5], [1, 2], [1, 3], [1, 4], [1, 5], 
+                    [2, 3], [2, 4], [2, 5], [3, 4], [3, 5], [4, 5], [1, 2, 3],
+                    [1, 3, 2], [1, 2, 4], [1, 4, 2], [1, 2, 5], [1, 5, 2], [1, 3, 4], 
+                    [1, 4, 3], [1, 3, 5], [1, 5, 3], [1, 4, 5], [1, 5, 4], [2, 3, 4], 
+                    [2, 4, 3], [2, 3, 5], [2, 5, 3], [2, 4, 5], [2, 5, 4], [3, 4, 5], 
+                    [3, 5, 4], [1, 2, 3, 4], [1, 2, 4, 3], [1, 3, 2, 4], [1, 3, 4, 2], 
+                    [1, 4, 2, 3], [1, 4, 3, 2], [1, 2, 3, 5], [1, 2, 5, 3], 
+                    [1, 3, 2, 5], [1, 3, 5, 2], [1, 5, 2, 3], [1, 5, 3, 2], 
+                    [1, 2, 4, 5], [1, 2, 5, 4], [1, 4, 2, 5], [1, 4, 5, 2], 
+                    [1, 5, 2, 4], [1, 5, 4, 2], [1, 3, 4, 5], [1, 3, 5, 4], 
+                    [1, 4, 3, 5], [1, 4, 5, 3], [1, 5, 3, 4], [1, 5, 4, 3], 
+                    [2, 3, 4, 5], [2, 3, 5, 4], [2, 4, 3, 5], [2, 4, 5, 3], 
+                    [2, 5, 3, 4], [2, 5, 4, 3], [1, 2, 3, 4, 5], [1, 2, 3, 5, 4], 
+                    [1, 2, 4, 3, 5], [1, 2, 4, 5, 3], [1, 2, 5, 3, 4], 
+                    [1, 2, 5, 4, 3], [1, 3, 2, 4, 5], [1, 3, 2, 5, 4], 
+                    [1, 3, 4, 2, 5], [1, 3, 4, 5, 2], [1, 3, 5, 2, 4], 
+                    [1, 3, 5, 4, 2], [1, 4, 2, 3, 5], [1, 4, 2, 5, 3], 
+                    [1, 4, 3, 2, 5], [1, 4, 3, 5, 2], [1, 4, 5, 2, 3], 
+                    [1, 4, 5, 3, 2], [1, 5, 2, 3, 4], [1, 5, 2, 4, 3], 
+                    [1, 5, 3, 2, 4], [1, 5, 3, 4, 2], [1, 5, 4, 2, 3], [1, 5, 4, 3, 2]]
+        self.assertEqual(expected, ur.gen_loops(n))
+
+    def test__loop2graph(self):
+        n = 3
+        e = [2, 3]
+        expected = {1: {}, 2: {3: 1}, 3: {2: 1}}
+        self.assertEqual(expected, ur.loop2graph(e, n))
+
+        n = 3
+        e = [1, 3, 2]
+        expected = {1: {3: 1}, 2: {1: 1}, 3: {2: 1}}
+        self.assertEqual(expected, ur.loop2graph(e, n))
+
+        n = 4
+        e = [1, 4, 3]
+        expected = {1: {4: 1}, 2: {}, 3: {1: 1}, 4: {3: 1}}
+        self.assertEqual(expected, ur.loop2graph(e, n))
+
+        n = 5
+        e = [1, 5, 4, 2]
+        expected = {1: {5: 1}, 2: {1: 1}, 3: {}, 4: {2: 1}, 5: {4: 1}}
+        self.assertEqual(expected, ur.loop2graph(e, n))
 
 
 class TestTraversalFunctions(unittest.TestCase):
